--- rthe.html
@@ -1,1 +1,0 @@
-<img src="./screenshots/app-preview-1.PNG" alt="" />
