+++ src/helper/util.js
@@ -1,28 +1,3 @@
-const db = require("../connection");
-
-const declareSqlQuery = (sqlQuery) => {
-  return db.query(sqlQuery, (err, result) => {
-    if (err) {
-      console.log(err);
-      return;
-    }
-
-    console.table(result);
-    db.end();
-  });
-};
-
-const displayChosenMySqlResultsTable = ({
-  homeMenu,
-  displayDepartments,
-  displayRoles,
-  displayEmployees,
-}) => {
-  if (homeMenu === "viewAllDepartments") return displayDepartments();
-  if (homeMenu === "viewAllRoles") return displayRoles();
-  if (homeMenu === "viewAllEmployees") return displayEmployees();
-};
-
 const verifyResponses = (input) => {
   return input ? true : "Please Enter Something";
 };
@@ -32,9 +7,37 @@ const verifyNumber = (input) => {
   return reg.test(input) || "Please Enter A Valid Number";
 };
 
+const generateDepartmentChoices = (departmentsFromDB) => {
+  return departmentsFromDB.map((department) => {
+    return {
+      name: department.name,
+      value: department.id,
+    };
+  });
+};
+
+const generateRoleChoices = (rolesFromDB) => {
+  return rolesFromDB.map((roles) => {
+    return {
+      name: roles.title,
+      value: roles.id,
+    };
+  });
+};
+
+const generateEmployeeChoices = (employeesFromDB) => {
+  return employeesFromDB.map((employee) => {
+    return {
+      name: `${employee.first_name} ${employee.last_name}: ${employee.name} `,
+      value: [employee.id, employee.department_id],
+    };
+  });
+};
+
 module.exports = {
-  declareSqlQuery,
-  displayChosenMySqlResultsTable,
   verifyNumber,
   verifyResponses,
+  generateDepartmentChoices,
+  generateRoleChoices,
+  generateEmployeeChoices,
 };
